% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.loadFiles}
\alias{.loadFiles}
\title{File loader
Higher order function to import BAM or Bowtie files.
Deals with wether type is \code{single} or \code{paired} and with the number of input
files}
\usage{
.loadFiles(singleLoad, pairedLoad)
}
\description{
File loader
Higher order function to import BAM or Bowtie files.
Deals with wether type is \code{single} or \code{paired} and with the number of input
files
}
