% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBAM.R
\name{.processStrand}
\alias{.processStrand}
\title{Process a given strand from a BAM file to read}
\usage{
.processStrand(strand, bam)
}
\arguments{
\item{strand}{either strand "+" or "-"}

\item{bam}{\code{data.frame} representing the BAM file as read by
\link[Rsamtools:scanBam]{Rsamtools::scanBam}}
}
\value{
\link[GenomicRanges:GRanges]{GenomicRanges::GRanges} object representing the reads in a given
strand
}
\description{
Process a given strand from a BAM file to read
}
