% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.unlist_as_integer}
\alias{.unlist_as_integer}
\title{Unlist an IRanges object into a vector}
\usage{
.unlist_as_integer(x)
}
\description{
Wrapper to internal function from the IRanges package. Avoids use of
\code{:::} and thus prevents a NOTE warning about the use of a non-exported
function
}
\author{
H. Pages, P. Aboyoun and M. Lawrence
}
