% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.vectorizedAll}
\alias{.vectorizedAll}
\title{Vectorized version of \code{all}}
\usage{
.vectorizedAll(...)
}
\arguments{
\item{...}{arbitraty amount of \code{logical} vectors, expected to have the same
length}
}
\value{
\code{logical} vector
}
\description{
Helper function that behaves as a vectorized version of the function \code{all}
}
