% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{nucleosome_htseq}
\alias{nucleosome_htseq}
\title{Example reads from high-troughtput sequencing nucleosome positioning
experiment}
\format{\code{RangedData} with the range of the reads and a data column with
the strand information.}
\source{
Publication pending
}
\description{
Few reads from paired-ended MNase-seq experiment in S.cerevisiae where
mononucleosomes were sequenced
}
\details{
This data is obtained from MNase digested nucleosomal DNA and sequenced with
Illumina platform. Paired-ended reads where mapped to SacCer1 genome using
Bowtie, and imported to R using the package \code{ShortRead} and paired ends
where merged into a single range.

Reads were sorted by chromosome and starting position and only a few reads
from the starting positions of chromosome 1 are presented.
}
\keyword{datasets}
