% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosomePatterns.R
\docType{methods}
\name{nucleosomePatterns}
\alias{nucleosomePatterns}
\alias{nucleosomePatterns,data.frame-method}
\alias{nucleosomePatterns,GRanges-method}
\alias{nucleosomePatterns,TxDb-method}
\title{Nucleosome patterns arround transcription start sites}
\usage{
nucleosomePatterns(genes, calls, window = 3000, p1.max.downstream = 20,
  open.thresh = 215, position = "tss", col.id, col.chrom, col.strand,
  col.pos)

\S4method{nucleosomePatterns}{data.frame}(genes, calls, window = 3000,
  p1.max.downstream = 20, open.thresh = 215, position = "tss",
  col.id = "name", col.chrom = "chrom", col.strand = "strand",
  col.pos = "start")

\S4method{nucleosomePatterns}{GRanges}(genes, calls, window = 3000,
  p1.max.downstream = 20, open.thresh = 215, position = "tss",
  col.id = "name", col.chrom = "seqnames", col.strand = "strand", col.pos)

\S4method{nucleosomePatterns}{TxDb}(genes, calls, window = 3000,
  p1.max.downstream = 20, open.thresh = 215, position = "tss",
  col.id = "tx_name", col.chrom = "seqnames", col.strand = "strand",
  col.pos)
}
\arguments{
\item{genes}{a \code{data.frame}, a \link[GenomicRanges:GRanges]{GenomicRanges::GRanges} or a
\link[GenomicFeatures:TxDb]{GenomicFeatures::TxDb} object representing the genes of the used organism}

\item{calls}{nucleosome calls performed with \code{nucleR}}

\item{window}{window in bp. around the start of the gene where the plus1 and
minus1 are searched. If no nucleosome is found in this window, it's
reported as \code{missing}.}

\item{p1.max.downstream}{the nucleosome plus1 can be upstream of the
transcription start site, but only this amount of base pairs.}

\item{open.thresh}{open/closed conformation threshold. Distances smaller
than this will be classified as \emph{close conformation}, while distances
bigger than this will be classified as \emph{open conformation}.}

\item{position}{whether our we are computing transcription start sites
(\code{tss}) or transcription termination sites (\code{tts}).}

\item{col.id, col.pos, col.strand, col.chrom}{names of the columns in the genes
\code{data.frame} containing the information corresponding to the gene id, the
position to compute (usually the transcription start site), the strand and
the chromosome for each gene.}
}
\value{
a \code{data.frame} with the following columns:
\itemize{
\item seqnames Name of the chromosome of the gene
\item id Name ID of the gene
\item p1.pos position of the plus 1 nucleosome
\item m1.pos position of the minus 1 nucleosome
\item dist distance in base pairs between \code{p1.pos} and \code{m1.pos}
\item descr descriptor of the conformation (W-open-W, W-close-F, +1_missing,
-1_missing, etc).
}
}
\description{
The transcription start site for each gene can be classified according to
the properties of the nucleosomes surrounding it (minus 1 and plus 1).
Those nucleosomes can be either \emph{well-positioned} (\code{W}), \emph{fuzzy} (\code{W}) or
\emph{missing}. And, when present, they can be at a distance less than a given
threshold (by default \code{215} bp.), which creates an \emph{close conformation} or
at a distance bigger than the threshold, which makes an \emph{open conformation}.
}
\author{
Ricard Illa \email{ricard.illa@irbbarcelona.org}
}
\keyword{mainp}
