% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBowtie.R
\name{readBowtie}
\alias{readBowtie}
\title{Import reads from a vector of Bowtie files}
\usage{
readBowtie(files, type = "paired")
}
\arguments{
\item{files}{List of input Bowtie files.}

\item{type}{Describes the type of reads. Values allowed are \code{single} for
single-ended reads and \code{paired} for pair-ended.}
}
\value{
List of \code{GRanges} containing the reads of each input BAM file.
}
\description{
This function allows to load reads from Bowtie files from both single and
paired-end commming from Next Generation Sequencing nucleosome mapping
experiments.
}
\author{
Ricard Illa \email{ricard.illa@irbbarcelona.org}
}
\keyword{file}
